requirements.txt
Flask==2.0.3
pandas==2.1.1
numpy==1.24.4
google-generativeai==0.3.1
google-cloud-storage==2.7.0
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.21.0
gunicorn==20.1.0
werkzeug==2.0.3